requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.10.0
django-heroku==0.3.1
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
idna==2.7
mccabe==0.6.1
Pillow==5.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.4
pytz==2018.5
requests==2.19.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.23
whitenoise==5.2.0
